refactor: move the type definitions to the suitable packages
