+++ packages/core/package.json
@@ -35,7 +35,6 @@
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@sunmao-ui/shared": "^0.2.0",
     "@swc/core": "^1.2.121",
     "@types/jest": "^26.0.23",
     "@types/json-schema": "^7.0.7",
