+++ packages/core/src/schema.ts
