+++ packages/runtime/src/components/_internal/ModuleRenderer.tsx
@@ -7,11 +7,12 @@ import {
   parseType,
   RuntimeApplication,
   RuntimeComponentSchema,
+  PropsAfterEvaled,
 } from '@sunmao-ui/core';
 import { ImplWrapper } from './ImplWrapper';
 import { watch } from '../../utils/watchReactivity';
 import { ImplementedRuntimeModule, UIServices } from '../../types';
-import { EventHandlerSpec, ModuleRenderSpec, PropsAfterEvaled } from '@sunmao-ui/shared';
+import { EventHandlerSpec, ModuleRenderSpec } from '@sunmao-ui/shared';
 import { resolveChildrenMap } from '../../utils/resolveChildrenMap';
 import { initStateAndMethod } from '../../utils/initStateAndMethod';
 import { ExpressionError } from '../../services/StateManager';
