+++ packages/runtime/src/types/index.ts
@@ -3,3 +3,4 @@ export * from './component';
 export * from './trait';
 export * from './module';
 export * from './utilMethod';
+export * from './spec';
