+++ packages/shared/src/types/index.ts
@@ -1,3 +1,1 @@
-export * from './spec';
 export * from './condition';
-export * from './schema';
